+++ packages/sdk/server-ai/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [0.11.3](https://github.com/launchdarkly/js-core/compare/server-sdk-ai-v0.11.2...server-sdk-ai-v0.11.3) (2025-08-29)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.16.1 to 2.16.2
+  * peerDependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.x to 2.16.2
+
+## [0.11.2](https://github.com/launchdarkly/js-core/compare/server-sdk-ai-v0.11.1...server-sdk-ai-v0.11.2) (2025-08-26)
+
+
+### Bug Fixes
+
+* **ai-sdk:** Remove Vercel mapping method from LD ([#911](https://github.com/launchdarkly/js-core/issues/911)) ([f71a457](https://github.com/launchdarkly/js-core/commit/f71a45774eb3fc0fd0ac6a93b1443843d14712d2))
+
+## [0.11.1](https://github.com/launchdarkly/js-core/compare/server-sdk-ai-v0.11.0...server-sdk-ai-v0.11.1) (2025-08-19)
+
+
+### Bug Fixes
+
+* Add usage tracking to config method ([#904](https://github.com/launchdarkly/js-core/issues/904)) ([7f0a54c](https://github.com/launchdarkly/js-core/commit/7f0a54c4d880c8bc784fee0d4a2bc9155e96c1b7))
+
 ## [0.11.0](https://github.com/launchdarkly/js-core/compare/server-sdk-ai-v0.10.1...server-sdk-ai-v0.11.0) (2025-08-01)